Move `wis2box.pubsub` to `wis2box.event`
Break modules into components. Rename components to better describe function
