+++ wis2box/event/subscribe.py
@@ -33,7 +33,7 @@ from wis2box.env import (BROKER_HOST, BROKER_PORT, BROKER_USERNAME,
                          BROKER_PASSWORD, STORAGE_SOURCE, STORAGE_ARCHIVE)
 from wis2box.handler import Handler
 from wis2box.plugin import load_plugin, PLUGINS
-from wis2box.pubsub.message import gcm
+from wis2box.event.messages.message import gcm
 
 LOGGER = logging.getLogger(__name__)
 
@@ -44,9 +44,7 @@ def handle(filepath):
         handler = Handler(filepath)
         if handler.handle():
             LOGGER.info('Data processed')
-            for plugin in handler.plugins:
-                for filepath in plugin.files():
-                    LOGGER.info(f'Public filepath: {filepath}')
+
     except ValueError as err:
         msg = f'handle() error: {err}'
         LOGGER.error(msg)
